renamed manifest reader files
