+++ discovery_engine/lib/src/infrastructure/assets/native/json_manifest_reader.dart
